+++ incrementationFunc.go
@@ -1,0 +1,16 @@
+package main
+
+type IncrementInput struct {
+	Key   string `json:"key"`
+	Value int    `json:"value"`
+}
+
+type IncrementResult struct {
+	Value int `json:"value"`
+}
+
+var IncrementMap = make(map[string]int)
+
+func IncrementingStructFunc(inp IncrementInput) IncrementResult {
+
+}
