+++ main.go
@@ -1,20 +1,5 @@
 package main
 
-type IncrementInput struct {
-	Key   string `json:"key"`
-	Value int    `json:"value"`
-}
-
-type IncrementResult struct {
-	Value int `json:"value"`
-}
-
-var IncrementMap = make(map[string]int)
-
-func IncrementingStructFunc(inp IncrementInput) IncrementResult {
-
-}
-
 func main() {
 
 }
